Move profiling out of examples
Profiling scripts will have their own home.
